Move post-processing off of the HPC, to the archival machine.
Rewrite wrapUpload.py, wrapReceive.py, and fillDbVasp.py
to move most of the processing off of the machine doing the
uploading.  This makes the upload process much faster, and the
researcher doesn't have to wait while we parse vasprun.xml files.

Originally we thought this would be a highly automated hands-off
process, and we could preprocess the vasprun.xml files on the HPC.
But it turns out that researchers need to review their results
and upload selectively by hand.  So the researcher may be waiting
for the upload.

Now all the processing of vasprun.xml occurs on the archival machine,
cid-dev.hpc.